Add a QapiSchema::parse fn
This will parse schema into a list of ordered Tokens representing
everything in the schema.json files except the whitespace. All comments
that begin a line are also captured in the return output.

The documentation parser/processor is coming next, but it wont ever
cause this main parser to *fail*. The parsing is done in a post
processing step, so if docs are missing it could be they got ignored and
just returned as normal comment blocks.