Complete schema file walking
All three schemas now parse correctly when starting from the
qapi-schema.json. The `include` paths are resolved relative to the
importated file. All in all this is pretty much done except for the
documentation parser for the comment blocks.

I would also like to stitch that documentation back into these struct
objects in a post-processing step for the parser.