Add trim_docstring
Admittedly this is a hack and I should have the functionality of what
this does in the parser-combinator. But the reference still holds. I
just lose (or rather gain) a little bit of context inside of nom errors
by not having nom aware of the extra trimming that is done.

Additionally start detecting if a member is optional or not based on `*`