Working unions are produced
Unions have been the worst thing to try to represent. I have landed on a
struct with a field called `branch: Option<Branch>`. This will get
flattened when serializing. For all the extra data a branch *may* hold
with a union, that gets pushed into a new enum `${unionname}Branch`.

The ser/de functions will need to account for this