Refactor to split the codegen out
I dont want the codegen in this crate, and I don't want the parser in
the codegen crate, so I am stripping this down to just an intermediate
representation. I will render the IR in the codegen crate. This should
be a stable representation going back to even the 0.15.0 qapi even as
the DSL changes.