Inject the base fields into the struct
This will bring the IR closer to the way it needs to be for #[repr(C)].
Instead of:
```
struct StructOne {
    /// doc comment one
    field_one: i64,
}
struct StructTwo {
    base: StructOne,
    /// doc comment two
    field_two: i64,
}
```
we now render in this new way:
```
struct StructOne {
    /// doc comment one
    field_one: i64,
}
struct StructTwo {
    /// doc comment one
    field_one: i64,
    /// doc comment two
    field_two: i64,
}
```

This does cause option duplication, but it matches what you would expect
in the C generated code. It also feels more ergonomical to use when
sending and constructing QMP Commands, in my opinion at least.