Only include name attr when name changes
Some names are disallowed by rust or may just be against the normal
convention for rust names. We still need to render the json to QMP with
the proper names. I could render this directly to a #[serde(rename = "")]
attribute, but I am trying to avoid direct serde usage in the redered
structs.

Removed add_name macro.